fix(#81): submit 버튼 미동작 오류 수정, UI 잘못 분리된 부분 수정
