+++ src/components/signUp/Nickname/style.ts
@@ -28,47 +28,3 @@ export const AlertMessage = styled.span<{color: "red" | "blue"}>`
     letter-spacing: 0.2px;
 `
 
-
-export const BrandIcon = styled.div<{
-    type: "kakao" | "google" | "naver" | "apple"
-}>`
-    width:16px;
-    height:16px;
-    border-radius:100%;
-
-    display:flex;
-    justify-content:center;
-    align-items:center;
-
-    svg{
-        width:16px;
-        height:16px;
-        path{
-            width:16px;
-            height:16px;
-        }
-    }
-
-    ${({type}) => type === "naver" && `svg{width:8px; height:8px;}`}
-
-    ${({type}) => {
-        switch(type) {
-            case "kakao":
-                return css`
-                    background-color: yellow;
-                `
-            case "google":
-                return css`
-                    background-color:#FFFFFF;
-                `
-            case "naver":
-                return css`
-                    background-color:#00BF18;
-                `
-            case "apple":
-                return css`
-                    background-color:#000000;
-                `
-        }
-    }}
-`
