+++ src/pages/auth/SignUpPage/style.ts
@@ -26,20 +26,64 @@ export const ButtonWrapper = styled.button`
     width:calc(100% - 40px);
     padding:17px 0px;
 
-    background-color: ${({ theme }) => theme.gray03};
+    background-color: ${({ theme }) => theme.main};
     
     cursor: pointer;
     border: none;
     border-radius: 30px;
     box-sizing:border-box;
 
-    color: ${({ theme }) => theme.white};
+    color: ${({ theme }) => theme.white };
     font-size: 16px;
     font-weight: 600;
     line-height: 22px;
 
     &:disabled {
-        background-color: ${({ theme }) => theme.main};
-        color: ${({ theme }) => theme.white };
+        background-color: ${({ theme }) => theme.gray03};
+        color: ${({ theme }) => theme.white};
     }
+`
+
+export const BrandIcon = styled.div<{
+    type: "kakao" | "google" | "naver" | "apple"
+}>`
+    width:16px;
+    height:16px;
+    border-radius:100%;
+
+    display:flex;
+    justify-content:center;
+    align-items:center;
+
+    svg{
+        width:16px;
+        height:16px;
+        path{
+            width:16px;
+            height:16px;
+        }
+    }
+
+    ${({type}) => type === "naver" && `svg{width:8px; height:8px;}`}
+
+    ${({type}) => {
+        switch(type) {
+            case "kakao":
+                return css`
+                    background-color: yellow;
+                `
+            case "google":
+                return css`
+                    background-color:#FFFFFF;
+                `
+            case "naver":
+                return css`
+                    background-color:#00BF18;
+                `
+            case "apple":
+                return css`
+                    background-color:#000000;
+                `
+        }
+    }}
 `
